+++ bgstally/windows/fleetcarrier.py
@@ -2,7 +2,7 @@ import tkinter as tk
 from functools import partial
 from tkinter import ttk
 
-from bgstally.constants import COLOUR_HEADING_1, FONT_HEADING_1, FONT_HEADING_2, FONT_TEXT, DiscordChannel, FleetCarrierItemType
+from bgstally.constants import COLOUR_HEADING_1, FONT_HEADING_1, FONT_HEADING_2, FONT_TEXT, DiscordFleetCarrier, DiscordChannel, FleetCarrierItemType
 from bgstally.debug import Debug
 from bgstally.fleetcarrier import FleetCarrier
 from bgstally.utils import _, __
@@ -19,7 +19,7 @@ class WindowFleetCarrier:
     def __init__(self, bgstally):
         self.bgstally = bgstally
 
-        self.toplevel:tk.Toplevel = None
+        self.toplevel: tk.Toplevel = None
 
 
     def show(self):
@@ -37,100 +37,117 @@ class WindowFleetCarrier:
         self.toplevel.iconphoto(False, self.bgstally.ui.image_logo_bgstally_32, self.bgstally.ui.image_logo_bgstally_16)
         self.toplevel.geometry("800x800")
 
-        container_frame:ttk.Frame = ttk.Frame(self.toplevel)
-        container_frame.pack(fill=tk.BOTH, expand=True)
+        frm_container: ttk.Frame = ttk.Frame(self.toplevel)
+        frm_container.pack(fill=tk.BOTH, expand=True)
 
-        ttk.Label(container_frame, text=_("System: {current_system} - Docking: {docking_access} - Notorious Allowed: {notorious}").format(current_system=fc.data.get('currentStarSystem', "Unknown"), docking_access=fc.human_format_dockingaccess(False), notorious=fc.human_format_notorious(False)), font=FONT_HEADING_1, foreground=COLOUR_HEADING_1).pack(anchor=tk.NW) # LANG: Label on carrier window
+        ttk.Label(frm_container, text=_("System: {current_system} - Docking: {docking_access} - Notorious Allowed: {notorious}").format(current_system=fc.data.get('currentStarSystem', "Unknown"), docking_access=fc.human_format_dockingaccess(False), notorious=fc.human_format_notorious(False)), font=FONT_HEADING_1, foreground=COLOUR_HEADING_1).pack(anchor=tk.NW) # LANG: Label on carrier window
 
         if not config.get_bool('capi_fleetcarrier'):
-            ttk.Label(container_frame, text=_("Some information cannot be updated. Enable Fleet Carrier CAPI Queries in File -> Settings -> Configuration"), foreground='#f00').pack(anchor=tk.NW) # LANG: Label on carrier window
+            ttk.Label(frm_container, text=_("Some information cannot be updated. Enable Fleet Carrier CAPI Queries in File -> Settings -> Configuration"), foreground='#f00').pack(anchor=tk.NW) # LANG: Label on carrier window
 
-        items_frame:ttk.Frame = ttk.Frame(container_frame)
-        items_frame.pack(fill=tk.BOTH, padx=5, pady=5, expand=True)
+        frm_items: ttk.Frame = ttk.Frame(frm_container)
+        frm_items.pack(fill=tk.BOTH, padx=5, pady=5, expand=True)
 
-        buttons_frame:ttk.Frame = ttk.Frame(container_frame)
-        buttons_frame.pack(fill=tk.X, padx=5, pady=5, side=tk.BOTTOM)
+        frm_buttons: ttk.Frame = ttk.Frame(frm_container)
+        frm_buttons.pack(fill=tk.X, padx=5, pady=5, side=tk.BOTTOM)
 
         current_row = 0
 
-        ttk.Label(items_frame, text=_("Selling Materials"), font=FONT_HEADING_2).grid(row=current_row, column=0, sticky=tk.W) # LANG: Label on carrier window
-        ttk.Label(items_frame, text=_("Buying Materials"), font=FONT_HEADING_2).grid(row=current_row, column=1, sticky=tk.W) # LANG: Label on carrier window
+        ttk.Label(frm_items, text=_("Selling Materials"), font=FONT_HEADING_2).grid(row=current_row, column=0, sticky=tk.W) # LANG: Label on carrier window
+        ttk.Label(frm_items, text=_("Buying Materials"), font=FONT_HEADING_2).grid(row=current_row, column=1, sticky=tk.W) # LANG: Label on carrier window
 
         current_row += 1
 
-        materials_selling_frame:ttk.Frame = ttk.Frame(items_frame)
-        materials_selling_text:TextPlus = TextPlus(materials_selling_frame, wrap=tk.WORD, height=1, font=FONT_TEXT)
-        materials_selling_scroll:tk.Scrollbar = tk.Scrollbar(materials_selling_frame, orient=tk.VERTICAL, command=materials_selling_text.yview)
-        materials_selling_text['yscrollcommand'] = materials_selling_scroll.set
-        materials_selling_scroll.pack(fill=tk.Y, side=tk.RIGHT)
-        materials_selling_text.pack(fill=tk.BOTH, side=tk.LEFT, expand=True)
+        frm_materials_selling: ttk.Frame = ttk.Frame(frm_items)
+        txt_materials_selling: TextPlus = TextPlus(frm_materials_selling, wrap=tk.WORD, height=1, font=FONT_TEXT)
+        sb_materials_selling: tk.Scrollbar = tk.Scrollbar(frm_materials_selling, orient=tk.VERTICAL, command=txt_materials_selling.yview)
+        txt_materials_selling['yscrollcommand'] = sb_materials_selling.set
+        sb_materials_selling.pack(fill=tk.Y, side=tk.RIGHT)
+        txt_materials_selling.pack(fill=tk.BOTH, side=tk.LEFT, expand=True)
 
-        materials_selling_text.insert(tk.INSERT, fc.get_items_plaintext(FleetCarrierItemType.MATERIALS_SELLING))
-        materials_selling_text.configure(state='disabled')
-        materials_selling_frame.grid(row=current_row, column=0, sticky=tk.NSEW)
+        txt_materials_selling.insert(tk.INSERT, fc.get_items_plaintext(FleetCarrierItemType.MATERIALS_SELLING))
+        txt_materials_selling.configure(state='disabled')
+        frm_materials_selling.grid(row=current_row, column=0, sticky=tk.NSEW)
 
-        materials_buying_frame:ttk.Frame = ttk.Frame(items_frame)
-        materials_buying_text:TextPlus = TextPlus(materials_buying_frame, wrap=tk.WORD, height=1, font=FONT_TEXT)
-        materials_buying_scroll:tk.Scrollbar = tk.Scrollbar(materials_buying_frame, orient=tk.VERTICAL, command=materials_buying_text.yview)
-        materials_buying_text['yscrollcommand'] = materials_buying_scroll.set
-        materials_buying_scroll.pack(fill=tk.Y, side=tk.RIGHT)
-        materials_buying_text.pack(fill=tk.BOTH, side=tk.LEFT, expand=True)
+        frm_materials_buying: ttk.Frame = ttk.Frame(frm_items)
+        txt_materials_buying: TextPlus = TextPlus(frm_materials_buying, wrap=tk.WORD, height=1, font=FONT_TEXT)
+        sb_materials_buying: tk.Scrollbar = tk.Scrollbar(frm_materials_buying, orient=tk.VERTICAL, command=txt_materials_buying.yview)
+        txt_materials_buying['yscrollcommand'] = sb_materials_buying.set
+        sb_materials_buying.pack(fill=tk.Y, side=tk.RIGHT)
+        txt_materials_buying.pack(fill=tk.BOTH, side=tk.LEFT, expand=True)
 
-        materials_buying_text.insert(tk.INSERT, fc.get_items_plaintext(FleetCarrierItemType.MATERIALS_BUYING))
-        materials_buying_text.configure(state='disabled')
-        materials_buying_frame.grid(row=current_row, column=1, sticky=tk.NSEW)
-
-        current_row += 1
-
-        ttk.Label(items_frame, text=_("Selling Commodities"), font=FONT_HEADING_2).grid(row=current_row, column=0, sticky=tk.W) # LANG: Label on carrier window
-        ttk.Label(items_frame, text=_("Buying Commodities"), font=FONT_HEADING_2).grid(row=current_row, column=1, sticky=tk.W) # LANG: Label on carrier window
+        txt_materials_buying.insert(tk.INSERT, fc.get_items_plaintext(FleetCarrierItemType.MATERIALS_BUYING))
+        txt_materials_buying.configure(state='disabled')
+        frm_materials_buying.grid(row=current_row, column=1, sticky=tk.NSEW)
 
         current_row += 1
 
-        commodities_selling_frame:ttk.Frame = ttk.Frame(items_frame)
-        commodities_selling_text:TextPlus = TextPlus(commodities_selling_frame, wrap=tk.WORD, height=1, font=FONT_TEXT)
-        commodities_selling_scroll:tk.Scrollbar = tk.Scrollbar(commodities_selling_frame, orient=tk.VERTICAL, command=commodities_selling_text.yview)
-        commodities_selling_text['yscrollcommand'] = commodities_selling_scroll.set
-        commodities_selling_scroll.pack(fill=tk.Y, side=tk.RIGHT)
-        commodities_selling_text.pack(fill=tk.BOTH, side=tk.LEFT, expand=True)
+        ttk.Label(frm_items, text=_("Selling Commodities"), font=FONT_HEADING_2).grid(row=current_row, column=0, sticky=tk.W) # LANG: Label on carrier window
+        ttk.Label(frm_items, text=_("Buying Commodities"), font=FONT_HEADING_2).grid(row=current_row, column=1, sticky=tk.W) # LANG: Label on carrier window
 
-        commodities_selling_text.insert(tk.INSERT, fc.get_items_plaintext(FleetCarrierItemType.COMMODITIES_SELLING))
-        commodities_selling_text.configure(state='disabled')
-        commodities_selling_frame.grid(row=current_row, column=0, sticky=tk.NSEW)
+        current_row += 1
 
-        commodities_buying_frame:ttk.Frame = ttk.Frame(items_frame)
-        commodities_buying_text:TextPlus = TextPlus(commodities_buying_frame, wrap=tk.WORD, height=1, font=FONT_TEXT)
-        commodities_buying_scroll:tk.Scrollbar = tk.Scrollbar(commodities_buying_frame, orient=tk.VERTICAL, command=commodities_buying_text.yview)
-        commodities_buying_text['yscrollcommand'] = commodities_buying_scroll.set
-        commodities_buying_scroll.pack(fill=tk.Y, side=tk.RIGHT)
-        commodities_buying_text.pack(fill=tk.BOTH, side=tk.LEFT, expand=True)
+        frm_commodities_selling: ttk.Frame = ttk.Frame(frm_items)
+        txt_commodities_selling: TextPlus = TextPlus(frm_commodities_selling, wrap=tk.WORD, height=1, font=FONT_TEXT)
+        sb_commodities_selling: tk.Scrollbar = tk.Scrollbar(frm_commodities_selling, orient=tk.VERTICAL, command=txt_commodities_selling.yview)
+        txt_commodities_selling['yscrollcommand'] = sb_commodities_selling.set
+        sb_commodities_selling.pack(fill=tk.Y, side=tk.RIGHT)
+        txt_commodities_selling.pack(fill=tk.BOTH, side=tk.LEFT, expand=True)
 
-        commodities_buying_text.insert(tk.INSERT, fc.get_items_plaintext(FleetCarrierItemType.COMMODITIES_BUYING))
-        commodities_buying_text.configure(state='disabled')
-        commodities_buying_frame.grid(row=current_row, column=1, sticky=tk.NSEW)
+        txt_commodities_selling.insert(tk.INSERT, fc.get_items_plaintext(FleetCarrierItemType.COMMODITIES_SELLING))
+        txt_commodities_selling.configure(state='disabled')
+        frm_commodities_selling.grid(row=current_row, column=0, sticky=tk.NSEW)
 
-        items_frame.columnconfigure(0, weight=1) # Make the first column fill available space
-        items_frame.columnconfigure(1, weight=1) # Make the second column fill available space
-        items_frame.rowconfigure(1, weight=1) # Make the materials text fill available space
-        items_frame.rowconfigure(3, weight=1) # Make the commodities text fill available space
+        frm_commodities_buying: ttk.Frame = ttk.Frame(frm_items)
+        txt_commodities_buying: TextPlus = TextPlus(frm_commodities_buying, wrap=tk.WORD, height=1, font=FONT_TEXT)
+        sb_commodities_buying: tk.Scrollbar = tk.Scrollbar(frm_commodities_buying, orient=tk.VERTICAL, command=txt_commodities_buying.yview)
+        txt_commodities_buying['yscrollcommand'] = sb_commodities_buying.set
+        sb_commodities_buying.pack(fill=tk.Y, side=tk.RIGHT)
+        txt_commodities_buying.pack(fill=tk.BOTH, side=tk.LEFT, expand=True)
 
-        self.copy_to_clipboard_button: tk.Button = tk.Button(buttons_frame, text=_("Copy to Clipboard"), command=partial(self._copy_to_clipboard, container_frame)) # LANG: Button label
-        self.copy_to_clipboard_button.pack(side=tk.LEFT, padx=5, pady=5)
+        txt_commodities_buying.insert(tk.INSERT, fc.get_items_plaintext(FleetCarrierItemType.COMMODITIES_BUYING))
+        txt_commodities_buying.configure(state='disabled')
+        frm_commodities_buying.grid(row=current_row, column=1, sticky=tk.NSEW)
 
-        self.post_button: tk.Button = tk.Button(buttons_frame, text=_("Post to Discord"), command=partial(self._post_to_discord)) # LANG: Button
-        self.post_button.pack(side=tk.RIGHT, padx=5, pady=5)
+        frm_items.columnconfigure(0, weight=1) # Make the first column fill available space
+        frm_items.columnconfigure(1, weight=1) # Make the second column fill available space
+        frm_items.rowconfigure(1, weight=1) # Make the materials text fill available space
+        frm_items.rowconfigure(3, weight=1) # Make the commodities text fill available space
+
+        self.btn_copy_to_clipboard: tk.Button = tk.Button(frm_buttons, text=_("Copy to Clipboard"), command=partial(self._copy_to_clipboard, frm_container)) # LANG: Button label
+        self.btn_copy_to_clipboard.pack(side=tk.LEFT, padx=5, pady=5)
+
+        self.btn_post_to_discord: tk.Button = tk.Button(frm_buttons, text=_("Post to Discord"), command=partial(self._post_to_discord)) # LANG: Button
+        self.btn_post_to_discord.pack(side=tk.RIGHT, padx=5, pady=5)
+        post_types: dict = {DiscordFleetCarrier.BOTH: _("Both Materials and Commodities"), # LANG: Dropdown menu on activity window
+                            DiscordFleetCarrier.MATERIALS: _("Materials Only"), # LANG: Dropdown menu on activity window
+                            DiscordFleetCarrier.COMMODITIES: _("Commodities Only")} # LANG: Dropdown menu on activity window
+        var_post_type: tk.StringVar = tk.StringVar(value=post_types.get(self.bgstally.state.DiscordFleetCarrier.get(), DiscordFleetCarrier.BOTH))
+        self.mnu_post_type: ttk.OptionMenu = ttk.OptionMenu(frm_buttons, var_post_type, var_post_type.get(),
+                                                            *post_types.values(),
+                                                            command=partial(self._post_type_selected, post_types), direction='above')
+        self.mnu_post_type.pack(side=tk.RIGHT, pady=5)
+        ttk.Label(frm_buttons, text=_("Activity to post:")).pack(side=tk.RIGHT, pady=5) # LANG: Label on activity window
+
         self._enable_post_button()
+
+
+    def _post_type_selected(self, post_types: dict, value: str):
+        """The user has changed the dropdown to choose the type of data to post
+        """
+        k: str = next(k for k, v in post_types.items() if v == value)
+        self.bgstally.state.DiscordFleetCarrier.set(k)
 
 
     def _post_to_discord(self):
         """
         Post Fleet Carrier materials list to Discord
         """
-        self.post_button.config(state=tk.DISABLED)
+        self.btn_post_to_discord.config(state=tk.DISABLED)
 
         fc: FleetCarrier = self.bgstally.fleet_carrier
 
-        title: str = __("Materials List for Carrier {carrier_name} in system: {system}", lang=self.bgstally.state.discord_lang).format(carrier_name=fc.name, system=fc.data.get('currentStarSystem', "Unknown")) # LANG: Discord fleet carrier title
+        title: str = __("Carrier {carrier_name}", lang=self.bgstally.state.discord_lang).format(carrier_name=fc.name) # LANG: Discord fleet carrier title
         description: str = self._get_as_text(fc, True)
 
         fields: list = []
@@ -140,7 +157,7 @@ class WindowFleetCarrier:
 
         self.bgstally.discord.post_embed(title, description, fields, None, DiscordChannel.FLEETCARRIER_MATERIALS, None)
 
-        self.post_button.after(5000, self._enable_post_button)
+        self.btn_post_to_discord.after(5000, self._enable_post_button)
 
 
     def _get_as_text(self, fc: FleetCarrier, short: bool = False) -> str:
@@ -156,20 +173,20 @@ class WindowFleetCarrier:
         text: str = ""
 
         if not short:
-            text += "**" + __("Materials List for Carrier {carrier_name}", lang=self.bgstally.state.discord_lang).format(carrier_name=fc.name) + "**\n\n" # LANG: Discord fleet carrier title
+            text += "**" + __("Carrier {carrier_name}", lang=self.bgstally.state.discord_lang).format(carrier_name=fc.name) + "**\n\n" # LANG: Discord fleet carrier title
 
         materials_selling: str = fc.get_items_plaintext(FleetCarrierItemType.MATERIALS_SELLING)
         materials_buying: str = fc.get_items_plaintext(FleetCarrierItemType.MATERIALS_BUYING)
         commodities_selling: str = fc.get_items_plaintext(FleetCarrierItemType.COMMODITIES_SELLING)
         commodities_buying: str = fc.get_items_plaintext(FleetCarrierItemType.COMMODITIES_BUYING)
 
-        if materials_selling != "":
+        if self.bgstally.state.DiscordFleetCarrier.get() != DiscordFleetCarrier.COMMODITIES and materials_selling != "":
             text += "**" + __("Selling Materials:", lang=self.bgstally.state.discord_lang) + f"**\n```css\n{materials_selling}```\n" # LANG: Discord fleet carrier section heading
-        if materials_buying != "":
+        if self.bgstally.state.DiscordFleetCarrier.get() != DiscordFleetCarrier.COMMODITIES and materials_buying != "":
             text += "**" + __("Buying Materials:", lang=self.bgstally.state.discord_lang) + f"**\n```css\n{materials_buying}```\n" # LANG: Discord fleet carrier section heading
-        if commodities_selling != "":
+        if self.bgstally.state.DiscordFleetCarrier.get() != DiscordFleetCarrier.MATERIALS and commodities_selling != "":
             text += "**" + __("Selling Commodities:", lang=self.bgstally.state.discord_lang) + f"**\n```css\n{commodities_selling}```\n" # LANG: Discord fleet carrier section heading
-        if commodities_buying != "":
+        if self.bgstally.state.DiscordFleetCarrier.get() != DiscordFleetCarrier.MATERIALS and commodities_buying != "":
             text += "**" + __("Buying Commodities:", lang=self.bgstally.state.discord_lang) + f"**\n```css\n{commodities_buying}```\n" # LANG: Discord fleet carrier section heading
 
         if not short:
@@ -184,7 +201,7 @@ class WindowFleetCarrier:
         """
         Re-enable the post to discord button if it should be enabled
         """
-        self.post_button.config(state=(tk.NORMAL if self._discord_button_available() else tk.DISABLED))
+        self.btn_post_to_discord.config(state=(tk.NORMAL if self._discord_button_available() else tk.DISABLED))
 
 
     def _discord_button_available(self) -> bool: