+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New Features:
 
 * Added ability to set the position of all in-game overlay panels. If you are finding you don't like where BGS-Tally positions the in-game information, or they overlap with other plugins, you can now override the default positions. If you would like to do this, take a look at the instructions in the `config\userconfig.template.ini` file.
+* Added options to only post your carrier materials, only your carrier commodities or both (defaults to both).
 
 ### Changes: